Finish fallback functionality to database and handle case if user is not connected to the internet.

+++ app/src/main/java/app/pixle/model/entity/item/Item.kt
@@ -10,5 +10,5 @@ data class Item(
 
     val category: String,
 
-    val difficulty: Int
+    val difficulty: Long
 )
